+++ src/DependencyInjection/ContaoBootstrapPanelExtension.php
@@ -14,7 +14,7 @@ final class ContaoBootstrapPanelExtension extends Extension
     /**
      * {@inheritdoc}
      */
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $loader = new XmlFileLoader(
             $container,